Remove volume sync
