+++ cluster-gateway/pkg/http/handlers_storage.go
@@ -44,24 +44,6 @@ func requireVolumeID(c *gin.Context) (string, bool) {
 	return id, true
 }
 
-func requireReplicaID(c *gin.Context) (string, bool) {
-	replicaID := c.Param("replica_id")
-	if replicaID == "" {
-		spec.JSONError(c, http.StatusBadRequest, spec.CodeBadRequest, "replica id is required")
-		return "", false
-	}
-	return replicaID, true
-}
-
-func requireConflictID(c *gin.Context) (string, bool) {
-	conflictID := c.Param("conflict_id")
-	if conflictID == "" {
-		spec.JSONError(c, http.StatusBadRequest, spec.CodeBadRequest, "conflict id is required")
-		return "", false
-	}
-	return conflictID, true
-}
-
 // createSandboxVolume creates a new sandbox volume
 func (s *Server) createSandboxVolume(c *gin.Context) {
 	c.Request.URL.Path = "/sandboxvolumes"
@@ -169,115 +151,5 @@ func (s *Server) deleteSandboxVolumeSnapshot(c *gin.Context) {
 		return
 	}
 	c.Request.URL.Path = "/sandboxvolumes/" + id + "/snapshots/" + snapshotID
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) upsertSyncReplica(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	replicaID, ok := requireReplicaID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/replicas/" + replicaID
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) getSyncReplica(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	replicaID, ok := requireReplicaID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/replicas/" + replicaID
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) appendSyncReplicaChanges(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	replicaID, ok := requireReplicaID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/replicas/" + replicaID + "/changes"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) updateSyncReplicaCursor(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	replicaID, ok := requireReplicaID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/replicas/" + replicaID + "/cursor"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) createSyncBootstrap(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/bootstrap"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) downloadSyncBootstrapArchive(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/bootstrap/archive"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) listSyncChanges(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/changes"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) downloadSyncReplayPayload(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/replay-payload"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) listSyncConflicts(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/conflicts"
-	s.proxyToStorageProxy(c)
-}
-
-func (s *Server) resolveSyncConflict(c *gin.Context) {
-	id, ok := requireVolumeID(c)
-	if !ok {
-		return
-	}
-	conflictID, ok := requireConflictID(c)
-	if !ok {
-		return
-	}
-	c.Request.URL.Path = "/sandboxvolumes/" + id + "/sync/conflicts/" + conflictID
 	s.proxyToStorageProxy(c)
 }
