+++ cluster-gateway/pkg/http/server.go
@@ -433,27 +433,6 @@ func (s *Server) setupRoutes() {
 				snapshots.DELETE("/:snapshot_id", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeDelete), s.deleteSandboxVolumeSnapshot)
 			}
 
-			sync := sandboxvolumes.Group("/:id/sync")
-			{
-				replicas := sync.Group("/replicas")
-				{
-					replicas.PUT("/:replica_id", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeWrite), s.upsertSyncReplica)
-					replicas.GET("/:replica_id", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeRead), s.getSyncReplica)
-					replicas.POST("/:replica_id/changes", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeWrite), s.appendSyncReplicaChanges)
-					replicas.PUT("/:replica_id/cursor", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeWrite), s.updateSyncReplicaCursor)
-				}
-
-				sync.POST("/bootstrap", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeWrite), s.createSyncBootstrap)
-				sync.GET("/bootstrap/archive", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeRead), s.downloadSyncBootstrapArchive)
-				sync.GET("/changes", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeRead), s.listSyncChanges)
-				sync.GET("/replay-payload", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeRead), s.downloadSyncReplayPayload)
-
-				conflicts := sync.Group("/conflicts")
-				{
-					conflicts.GET("", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeRead), s.listSyncConflicts)
-					conflicts.PUT("/:conflict_id", s.authMiddleware.RequirePermission(gatewayauthn.PermSandboxVolumeWrite), s.resolveSyncConflict)
-				}
-			}
 		}
 	}
