+++ docs/manifest.json
@@ -162,10 +162,6 @@
         {
           "slug": "fork",
           "title": "Fork"
-        },
-        {
-          "slug": "sync",
-          "title": "Sync"
         }
       ]
     },
