+++ docs/volume/page.mdx
@@ -297,14 +297,6 @@ This means:
   </LinkCard>
 
   <LinkCard
-    title="Volume Sync"
-    href="/docs/volume/sync"
-    cta="Learn More"
-  >
-    Keep one workspace in sync across local machines and sandboxes
-  </LinkCard>
-
-  <LinkCard
     title="Sandbox"
     href="/docs/sandbox"
     cta="Learn More"
